@@ -16,11 +16,10 @@
 package com.google.samples.gridtopager;
 
 import android.os.Bundle;
-import android.os.PersistableBundle;
-import android.support.v4.app.FragmentManager;
-import android.support.v7.app.AppCompatActivity;
+
+import androidx.fragment.app.FragmentManager;
+import androidx.appcompat.app.AppCompatActivity;
 import com.google.samples.gridtopager.fragment.GridFragment;
-import com.google.samples.gridtopager.R;
 
 /**
  * Grid to pager app's main activity.
