@@ -19,9 +19,9 @@ package com.google.samples.gridtopager.adapter;
 import static com.google.samples.gridtopager.adapter.ImageData.IMAGE_DRAWABLES;
 
 import android.graphics.drawable.Drawable;
-import android.support.annotation.Nullable;
-import android.support.v4.app.Fragment;
-import android.support.v7.widget.RecyclerView;
+import androidx.annotation.Nullable;
+import androidx.fragment.app.Fragment;
+import androidx.recyclerview.widget.RecyclerView;
 import android.transition.TransitionSet;
 import android.view.LayoutInflater;
 import android.view.View;
